+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - decoration components
 - nav component styling and intersection api
 - CardsColumns.js and Card.js
+- footer copyright
 - Header.js component
 
 ### Fixed